+++ R/dbSource.R
@@ -1,3 +1,18 @@
+# Copyright 2024 DARWIN EU®
+#
+# This file is part of CDMConnector
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 #' Create a source for a cdm in a database.
 #'
@@ -6,7 +21,6 @@
 #' write access to it.
 #'
 #' @export
-#'
 dbSource <- function(con, writeSchema) {
   # initial checks
   if (methods::is(con, "Pool")) {
@@ -17,7 +31,7 @@ dbSource <- function(con, writeSchema) {
   }
   if (methods::is(con, "DatabaseConnectorConnection")) {
     cli::cli_warn(
-      "Not all functionality is supported when DatabaseConnector as your
+      "Not all functionality is supported when DatabaseConnector is your
       database driver! Some issues may occur."
     )
   }
@@ -177,5 +191,74 @@ insertFromSource.db_cdm <- function(cdm, value) {
     table = value, src = attr(cdm, "cdm_source"), name = remoteName
   )
   return(value)
+}
+
+#' @export
+#' @importFrom omopgenerics cdmTableFromSource
+cdmTableFromSource.db_cdm <- function(src, value) {
+  if (inherits(value, "data.frame")) {
+    cli::cli_abort(
+      "To insert a local table to a cdm_reference object use insertTable
+      function."
+    )
+  }
+  if (!inherits(value, "tbl_lazy")) {
+    cli::cli_abort(
+      "Can't assign an object of class: {paste0(class(value), collapse = ', ')}
+      to a db_con cdm_reference object."
+    )
+  }
+  con <- attr(src, "dbcon")
+  schema <- attr(src, "write_schema")
+  if (!identical(con, dbplyr::remote_con(value))) {
+    cli::cli_abort(
+      "The cdm object and the table have different connection sources."
+    )
+  }
+  remoteName <- dbplyr::remote_name(value)
+  if ("dbplyr" == substr(remoteName, 1, 6)) {
+    remoteName <- NA_character_
+  } else if ("prefix" %in% names(schema)) {
+    prefix <- schema["prefix"] |> unname()
+    if (substr(remoteName, 1, nchar(prefix)) == prefix) {
+      remoteName <- substr(remoteName, nchar(prefix) + 1, nchar(remoteName))
+    }
+  }
+  value <- omopgenerics::newCdmTable(
+    table = value, src = src, name = remoteName
+  )
+  return(value)
+}
+
+#' @export
+listSourceTables.db_cdm <- function(cdm) {
+  listTables(con = attr(cdm, "dbcon"), schema = attr(cdm, "write_schema"))
+}
+
+#' @export
+dropSourceTable.db_cdm <- function(cdm, name) {
+  # initial checks
+  schema <- attr(cdm, "write_schema")
+  con <- attr(cdm, "dbcon")
+  checkmate::assertTRUE(DBI::dbIsValid(con))
+
+  # drop tables
+  for (i in seq_along(name)) {
+    DBI::dbRemoveTable(conn = con, name = inSchema(
+      schema = schema, table = name[i], dbms = dbms(con)
+    ))
+  }
+
+  return(cdm)
+}
+
+#' @export
+readSourceTable.db_cdm <- function(cdm, name) {
+  con <- attr(cdm, "dbcon")
+  schema <- attr(cdm, "write_schema")
+  fullName <- inSchema(schema = schema, table = name, dbms = dbms(con))
+  dplyr::tbl(src = con, fullName) |>
+    dplyr::rename_all(tolower) |>
+    omopgenerics::newCdmTable(src = cdm, name = tolower(name))
 }