+++ R/reexports-omopgenerics.R
@@ -22,4 +22,23 @@ omopgenerics::insertTable
 #' @export
 omopgenerics::dropTable
 
+#' @importFrom omopgenerics readSourceTable
+#' @export
+omopgenerics::readSourceTable
+
+#' @importFrom omopgenerics dropSourceTable
+#' @export
+omopgenerics::dropSourceTable
+
+#' @importFrom omopgenerics listSourceTables
+#' @export
+omopgenerics::listSourceTables
+
+#' @importFrom omopgenerics bind
+#' @export
+omopgenerics::bind
+
+#' @importFrom omopgenerics cdmVersion
+#' @export
+omopgenerics::cdmVersion