+++ tests/testthat/test-local_cdm.R
@@ -9,12 +9,12 @@ test_that("assertTables works with local cdms", {
     con = con, cdm_name = "eunomia", cdm_schema = "main", write_schema = "main"
   )
 
-  expect_equal(version(cdm), "5.3")
+  expect_equal(cdmVersion(cdm), "5.3")
   expect_error(assertTables(cdm, "concept"), NA)
 
   # local r
   cdm_r <- dplyr::collect(cdm)
-  expect_equal(version(cdm_r), "5.3")
+  expect_equal(cdmVersion(cdm_r), "5.3")
   assertTables(cdm_r, "concept")
 
   # arrow